Structural changes to S3 upload + legacy/non-legacy-split wip
