+++ .devcontainer/Dockerfile
@@ -21,7 +21,7 @@ RUN if [ "${INSTALL_NODE}" = "true" ]; then su vscode -c "umask 0002 && . /usr/l
 RUN apt update \
     && export DEBIAN_FRONTEND=noninteractive \
     && apt -y install --no-install-recommends postgresql-client-common \
-    && apt-get install --fix-missing libcurl4-openssl-dev libssl-dev
+    && apt -y install --fix-missing libcurl4-openssl-dev libssl-dev
 
 # Copy install and launcher script to bin:
 COPY ./dev_install /bin
